% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anime.R
\name{get_anime_details}
\alias{get_anime_details}
\title{Get Anime Details}
\usage{
get_anime_details(
  anime_id,
  fields = c("id", "title", "main_picture"),
  client_id
)
}
\arguments{
\item{anime_id}{anime ID as a string}

\item{fields}{list of desired fields to include;
acceptable values include: \cr
id, title,main_picture, alternative_titles, start_date,
end_date, synopsis, mean, rank, popularity, num_list_users, num_scoring_users,
nsfw, created_at, updated_at, media_type, status, genres, my_list_status,
num_episodes, start_season, broadcast, source, average_episode_duration, rating,
pictures, background, related_anime, related_manga, recommendations, studios, statistics \cr \cr
use c("all") to include all fields; \cr
by default, "id", "title", and "main_picture" are always returned}

\item{client_id}{your MAL Client ID; \href{https://myanimelist.net/blog.php?eid=835707}{instructions to get a MAL Client ID}}
}
\description{
This function is like viewing an anime's page on myanimelist.net.
The ID can be obtained from either using the \code{\link{get_anime_search}} function or the url
(for example, the ID of \href{https://myanimelist.net/anime/11061/Hunter_x_Hunter_2011}{Hunter x Hunter (2011)} is 11061).
}
\examples{
get_anime_details(anime_id = "11061", client_id = "your client id")

get_anime_details(anime_id = "11061", fields = c("all"), client_id = "your client id")

get_anime_details(anime_id = "11061",
                    fields = c("related_anime", "num_episodes"),
                 client_id = "your client id")
}
