% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anime.R
\name{get_anime_ranking}
\alias{get_anime_ranking}
\title{Get Anime Ranking}
\usage{
get_anime_ranking(
  ranking_type = "all",
  limit = 100,
  offset = 0,
  fields = c("id", "title", "main_picture"),
  client_id
)
}
\arguments{
\item{ranking_type}{ranking category; acceptable values include: \cr
all, airing, upcoming, tv, ova, movie, special, bypopularity, favorite \cr \cr
defaults to "all"}

\item{limit}{limits number of results; defaults to 100}

\item{offset}{number of results to skip; defaults to 0}

\item{fields}{list of desired fields to include;
acceptable values include: \cr
id, title,main_picture, alternative_titles, start_date,
end_date, synopsis, mean, rank, popularity, num_list_users, num_scoring_users,
nsfw, created_at, updated_at, media_type, status, genres, my_list_status,
num_episodes, start_season, broadcast, source, average_episode_duration, rating,
pictures, background, related_anime, related_manga, recommendations, studios, statistics \cr \cr
use c("all") to include all fields; \cr
by default, "id", "title", and "main_picture" are always returned}

\item{client_id}{your MAL Client ID; \href{https://myanimelist.net/blog.php?eid=835707}{instructions to get a MAL Client ID}}
}
\description{
This function is like viewing the \href{https://myanimelist.net/topanime.php}{Top Anime} section on myanimelist.net.
}
\examples{
get_anime_ranking(client_id = "your client id")

get_anime_ranking(ranking_type = "airing", fields = c("all"), client_id = "your client id")

get_anime_ranking(ranking_type = "upcoming",
                         limit = 10,
                        offset = 0,
                        fields = c("synopsis", "genres"),
                     client_id = "your client id")
}
