% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anime.R
\name{get_anime_search}
\alias{get_anime_search}
\title{Search for Anime}
\usage{
get_anime_search(
  search_term,
  limit = 100,
  offset = 0,
  fields = c("id", "title", "main_picture"),
  client_id
)
}
\arguments{
\item{search_term}{search anime by this term}

\item{limit}{limits number of results; defaults to 100}

\item{offset}{number of results to skip; defaults to 0}

\item{fields}{list of desired fields to include;
acceptable values include: \cr
id, title,main_picture, alternative_titles, start_date,
end_date, synopsis, mean, rank, popularity, num_list_users, num_scoring_users,
nsfw, created_at, updated_at, media_type, status, genres, my_list_status,
num_episodes, start_season, broadcast, source, average_episode_duration, rating,
pictures, background, related_anime, related_manga, recommendations, studios, statistics \cr \cr
use c("all") to include all fields; \cr
by default, "id", "title", and "main_picture" are always returned}

\item{client_id}{your MAL Client ID; \href{https://myanimelist.net/blog.php?eid=835707}{instructions to get a MAL Client ID}}
}
\description{
This function is like using the search bar for anime on myanimelist.net.
}
\examples{
get_anime_search(search_term = "one punch", client_id = "your client id")

get_anime_search(search_term = "one punch", fields = c("all"), client_id = "your client id")

get_anime_search(search_term = "one punch",
                       limit = 10,
                      offset = 0,
                      fields = c("mean", "genres"),
                   client_id = "your client id")
}
