% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manga.R
\name{get_manga_details}
\alias{get_manga_details}
\title{Get Manga Details}
\usage{
get_manga_details(
  manga_id,
  fields = c("id", "title", "main_picture"),
  client_id
)
}
\arguments{
\item{manga_id}{manga ID as a string}

\item{fields}{list of desired fields to include;
acceptable values include: \cr
id, title, main_picture, alternative_titles, start_date, end_date, synopsis,
mean, rank, popularity, num_list_users, num_scoring_users, nsfw, created_at,
updated_at, media_type, status, genres, my_list_status, num_volumes, num_chapters,
authors\{first_name,last_name\}, pictures, background, related_anime, related_manga,
recommendations, serialization \cr \cr
use c("all") to include all fields; \cr
by default, "id", "title", and "main_picture" are always returned}

\item{client_id}{your MAL Client ID; \href{https://myanimelist.net/blog.php?eid=835707}{instructions to get a MAL Client ID}}
}
\description{
This function is like viewing a manga's page on myanimelist.net.
The ID can be obtained from either using the \code{\link{get_manga_search}} function or the url
(for example, the ID of \href{https://myanimelist.net/manga/51/Slam_Dunk}{Slam Dunk} is 51).
}
\examples{
get_manga_details(anime_id = "51", client_id = "your client id")

get_manga_details(anime_id = "51", fields = c("all"), client_id = "your client id")

get_manga_details(anime_id = "51",
                    fields = c("synopsis", "authors{first_name,last_name}"),
                 client_id = "your client id")

}
