% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anime.R
\name{get_seasonal_anime}
\alias{get_seasonal_anime}
\title{Get Seasonal Anime}
\usage{
get_seasonal_anime(
  year = format(Sys.Date(), "\%Y"),
  season = get_current_season(),
  sort = "anime_score",
  limit = 100,
  offset = 0,
  fields = c("id", "title", "main_picture"),
  client_id
)
}
\arguments{
\item{year}{desired year as string or numeric; \cr
defaults to the current year}

\item{season}{desired season as string; acceptable values include: winter, spring, summer, fall \cr
defaults to the current season}

\item{sort}{how to order results (descending); acceptable values include: anime_score, anime_num_list_users \cr
defaults to "anime_score"}

\item{limit}{limits number of results; defaults to 100}

\item{offset}{number of results to skip; defaults to 0}

\item{fields}{list of desired fields to include;
acceptable values include: \cr
id, title,main_picture, alternative_titles, start_date,
end_date, synopsis, mean, rank, popularity, num_list_users, num_scoring_users,
nsfw, created_at, updated_at, media_type, status, genres, my_list_status,
num_episodes, start_season, broadcast, source, average_episode_duration, rating,
pictures, background, related_anime, related_manga, recommendations, studios, statistics \cr \cr
use c("all") to include all fields; \cr
by default, "id", "title", and "main_picture" are always returned}

\item{client_id}{your MAL Client ID; \href{https://myanimelist.net/blog.php?eid=835707}{instructions to get a MAL Client ID}}
}
\description{
This function is like viewing the \href{https://myanimelist.net/anime/season}{Seasonal Anime} section on myanimelist.net.
}
\examples{
get_seasonal_anime(client_id = "your client id")

get_seasonal_anime(year = "2022",
                 season = "winter",
                 fields = c("all"),
              client_id = "your client id")

get_seasonal_anime(year = "2022",
                  season = "winter",
                    sort = "anime_num_list_users",
                   limit = 10,
                  offset = 0,
                  fields = c("all"),
               client_id = "your client id")
}
